### 2018-06-18 de5b690
removed list items that don't even exist
