### 2018-06-25 #778 @arfon
@viktor76525 - it looks like the whitespace isn't quite right in your `paper.md` - could you make sure all of the YAML keys are lined up correctly?
