### 2018-06-25 a2ae434
removed table of contents in readme.md
