### 2018-06-25 #778 @viktor76525
@arfon I'm not sure what you mean, I thought the format of `paper.md` was GitHub markdown. I think the issue may have been with the table of contents and I have removed it as it was unnecessary anyway. Please let me know if there are further changes required.
