### 2018-06-26 #778 @arfon
I've fixed this for you here. YAML is whitespace sensitive and your fields for authors and affiliations had inconsistent indentations: https://github.com/ucgmsim/GMSimViz/pull/1
