### 2018-06-26 #778 @viktor76525
Thank you. Sorry, I mixed up `paper.md` with `readme.md`.
