### 2018-06-28 #778 @lheagy
👋 Hi @arfon and @viktor76525: I would be happy to be the editor on this submission
