### 2018-07-02 #778 @leouieda
@lheagy I was actually taking a look at this and would like to review it. I might take a while, though, with Scipy coming up next week. Would that be OK?

I think @seisman and @joa-quim might be interested in reviewing this as well.
