### 2018-07-03 #778 @lheagy
Thanks @leouieda and @viktor76525! @leouieda, getting the review in after SciPy is fine - enjoy the conference
