### 2018-07-03 #778 @joa-quim
Well, I wouldn't mind help the review but the condition of "Linux only" is not for me. I'm a Windows guy. And also, Python is not a language that I practice. So, I don't think I would be very helpful.
