### 2018-07-04 #778 @hugoledoux
Is it Linux only? I don't know all these packages but I see no reason why it wouldn't work on macOS (Windows too I guess).

I can review this, and see if that works on macOS
