### 2018-07-04 #808 @whedon
[REVIEW]: GMSimViz
**Submitting author:** @viktor76525 (<a href="http://orcid.org/0000-0001-5442-2960">Viktor</a>)
**Repository:** <a href="https://github.com/ucgmsim/GMSimViz" target ="_blank">https://github.com/ucgmsim/GMSimViz</a>
**Version:** 1.3

- [x] **Version:** 1.3
- [x] **Authorship:** Has the submitting author (@￼viktor76525) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

- [x] **Version:** 1.3
- [x] **Authorship:** Has the submitting author (@￼viktor76525) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?
