### 2018-07-04 #808 @lheagy
👋 Thanks @leouieda and @hugoledoux for being willing to review! This issue has a checklist for each of you in the main issue thread to help guide the review. Please let me know if you have any questions or if I can help clarify anything!
