### 2018-07-04 ucgmsim/GMSimViz#3 @lheagy
JOSS paper ORCID's
The ORCID's provided in the JOSS paper submitted in openjournals/joss-reviews#808 are not valid ORICD's (we just started checking this: openjournals/whedon#29, so it may have compiled previously). @viktor76525, you and your co-authors will need to obtain valid ORCID ID's for the submission.

It is free and quick to register: https://orcid.org/register
