+++ paper/paper.md
@@ -6,7 +6,7 @@ tags:
   - automated workflow
 authors:
   - name: Viktor Polak
-    orcid: 0000-0000-0000-0000
+    orcid: 0000-0001-5442-2960
     affiliation: "1"
   - name: Melody Zhu
     orcid: 0000-0000-0000-0000
