### 2018-07-05 ucgmsim/GMSimViz#3 @viktor76525
@lheagy should be all sorted now
