### 2018-07-26 #808 @hugoledoux
# version

there is no release

# installation

I accepted thinking, perhaps foolishly, that I could just install it under macOS. But after trying to install a few libraries, I just saw this: https://github.com/ucgmsim/GMSimViz#other-systems

Is this because you've never tried, or because you know it's impossible?

Also, I could install most of the packages, but sometimes it's difficult which one you mean. Like with MPI, under macOS one can do this and that works:

  1. `brew install open-mpi`
  2. `pip install mpi4py`

But it would be nice to know which python package (URL in pypi) is necessary.
For FFMpeg, there are many of them, which ones do you use?

Also stating Python >2.6 is false, I started with the current version of Python (3.6.5), and noticed that you mean Python2 only... Python2 is deprecated from next year I believe, why still use it?

So I'd like to know if possible under macOS before.
But I have the feeling that if only Linux, then can it be accepted?
What are the rules of JOSS?
@lheagy : can you help us here?
Also, even if only Linux, the installation of the dependencies is very "hacky", ie one must fiddle a lot... Is it possible to streamline the installation?
