### 2018-07-27 #808 @lheagy
Hi @hugoledoux, thanks for your comments. To answer your question, we can accept platform-specific packages. If it runs only on Linux, that is okay. However, if having a workflow to install on other systems (e.g. mac) is not overly challenging, I would encourage that to be added to the docs (even if there is a disclaimer that the workflow is a bit hacky) - This would increase the potential audience and user-base of the package.

Regarding Python 2 vs Python 3 - we can accept a python 2 specific package. In the docs or issues on the repository, any plans to upgrade to python 3 (or not to) should be explained. The community is moving to Python 3 and Python 2 will be losing support, so at least having a plan to migrate over would be important for longevity.

I hope this helps.
