### 2018-08-10 #808 @lheagy
👋 Hi @leouieda, @hugoledoux, @viktor76525: Just checking in to see how things are going
