### 2018-08-12 #808 @viktor76525
Sorry for the delay in response. I was away for a few weeks but am now back.

@hugoledoux is right, there should be no reason why it doesn't work on OSX, we had no machines to test on, unsupported means we cannot give instructions for those who are not confident with their environment. Technically the instructions under Linux are for apt based distributions.

Python >2.6 is for Python 2, not 3. ie 2.6<=python<3, if this is an issue I can go through and make necessary changes. The reason for development was organisational resistance which I agree is nonsense.

MPI is required in Python, therefore mpi4py is required, I thought open-mpi is a dependency of mpi4py. I can go through and check if it is not clear.

Any FFMPEG installation should be fine, I doubt you can find a recent build that doesn't work. You are right that the description is too detailed as it includes default-on compile time options and therefore will confuse everyone. On some supercomputing environments it is difficult to build a complete ffmpeg so in this case it was important to know the absolute minimum requirements. These can be removed from the instructions or made clearer if you wish.
