### 2018-08-22 #808 @hugoledoux
Right now I can't install the software so difficult for me to review it...

So yes if you could check what is necessary (the URL of the pacakge used) then I can give it another go.
