### 2018-08-28 ec3cd00
gmsimviz python 3 compatibility
