### 2018-08-28 3e1175c
streamlined installation
