### 2018-08-28 8147fb6
more notes/clearup
