### 2018-08-28 #808 @viktor76525
Thanks for the suggestions @hugoledoux
GMSimViz now works with Python 3 (3.6 tested with the sample dataset)
The installation has been streamlined (no more manual editing of the text file)
I will quickly add the URLs for the Python packages listed but note that they require the backends too, these should be installed automatically if using a package manager.
