### 2018-08-28 1bd5c9b
added URLs for dependencies
