### 2018-09-04 ucgmsim/GMSimViz#4 @leouieda
Update and fill out the setup.py (JOSS review)
Why is the `setup.py` for the `qcore` package? The repository and everything in there refers to `GMSimViz` instead of `qcore`. I would recommend renaming `qcore` to `gmsimviz` and making the CLI application a module in this package. See https://python-packaging.readthedocs.io/en/latest/command-line-scripts.html

Also, the `setup.py` file is not complete. This makes it impossible to install GMSimViz using standard tools like pip. Doing all this would allow your users to `pip install gmsimviz` to get latest release.

The [Python Packaging Guide](https://packaging.python.org/) is a good resource to get started. The [Python Packaging Tutorial](https://python-packaging.readthedocs.io/en/latest/index.html) is also very good.

openjournals/joss-reviews#808
