### 2018-09-04 ucgmsim/GMSimViz#5 @leouieda
Make an official release (JOSS review)
The software needs an official release to mark which exact commit is "1.0.0". You can do this by adding a git tag or online at https://github.com/ucgmsim/GMSimViz/releases. This way, when you say GMSimViz 1.0.0, we know exactly what it means.

Ideally, there should also be a release on PyPI. I see that you have a `setup.py` file but it seems to be for the "qcore" package, which is a bit confusing. See #4 for some comments on that.

openjournals/joss-reviews#808
