### 2018-09-04 #808 @leouieda
Hi @lheagy and @viktor76525, sorry for the long delay. I'll post my comments here and also open issues in the repository.
