### 2018-09-04 ucgmsim/GMSimViz#6 @leouieda
Clearer install instruction for the package (JOSS review)
The install instructions show how to install the dependencies but not the GMSimViz software it self. I'm assuming that it is implied that the user should copy GMSimViz around to where the data are (or the other way around). If this is the case, there should be instructions for that. This is not ideal, though. Many users won't know where to download the source code, how to git clone or get a zip archive from Github, etc. There should instructions for this if that's the preferred install method.

A much better option would be make the GMSimViz package installable. See #4. I highly recommend doing this instead.

openjournals/joss-reviews#808
