### 2018-09-04 ucgmsim/GMSimViz#7 @leouieda
Show the expected outcome of the example (JOSS review)
There is an example usage section in the README but it doesn't show the expected outcome. So a user has no way of know if they got the correct result. It would be best to include a screenshot, movie, or youtube link of the final product.

openjournals/joss-reviews#808
