### 2018-09-04 ucgmsim/GMSimViz#8 @leouieda
Create a smaller first example (JOSS review)
The example in the README is reported to take 45min to execute. This is a long time to wait if you're just trying the software out and want to see if it can be used for your project or experiment with the options. A first example should ideally be quick to run and easy to tweak for new users. It doesn't have to show every single element of the software. It can be complemented by a more detailed and longer running example.

openjournals/joss-reviews#808
