### 2018-09-04 ucgmsim/GMSimViz#9 @leouieda
Full documentation of the scripts options (JOSS review)
The only documentation provided in the project is the brief description of the example usage. The command-line help for the script also only briefly describes each option. There should be a full description of each option and what kinds of values they can take. Ideally, there could be an example showcasing what each option does to the final product.

openjournals/joss-reviews#808
