### 2018-09-04 ucgmsim/GMSimViz#10 @leouieda
Refactor the code for testing (JOSS review)
It would be very hard to test the CLI program the way it is written right now (everything in massive functions in a single file). The reason is that it is impossible to test small portions of the code in isolation. These are known as unit tests and are extremely useful when refactoring your code because you immediately know if you broke something (the tests will fail).

I recommend breaking up the `GMSimViz.py` module into several parts. A good rule of thumb is to have functions with at most ~50 lines of code. Then you can test them individually in isolation. This is related to #4.

For a great guide on getting started with testing and package design, see https://docs.python-guide.org/.

openjournals/joss-reviews#808
