### 2018-09-04 ucgmsim/GMSimViz#11 @leouieda
Create community guidelines (JOSS review)
Currently, there is no information for people wishing to contribute to the project. It is the current standard for open-source projects to have:

* A `CONTRIBUTING.md` file with information about how you want people to send contributions to your project. I have a standard guide that I use and tweak for most of my projects here: https://github.com/fatiando/contributing
* A `CODE_OF_CONDUCT.md` file with a code of conduct for the project. You don't have to write your own. There is a great one at https://www.contributor-covenant.org/, which is what I use.

openjournals/joss-reviews#808
