### 2018-09-04 ucgmsim/GMSimViz#12 @leouieda
Properly cite GMT (JOSS review)
The GMT website (http://gmt.soest.hawaii.edu/projects/gmt/wiki/References) indicates that the latest citation for GMT is the 2013 EOS article (https://doi.org/10.1002/2013EO450001). You can get a Bibtex for that reference at https://agupubs.onlinelibrary.wiley.com/action/showCitFormats?doi=10.1002%2F2013EO450001

openjournals/joss-reviews#808
