### 2018-09-04 #808 @leouieda
@lheagy and @viktor76525, I've gone through the review checklist and some things are missing or need improvement. This is a summary of my comments. See the related issues for more details.

- [ ] Version: There is no released version of this software on Github or PyPI. See ucgmsim/GMSimViz#4 and ucgmsim/GMSimViz#5
- [ ] The install instructions don't mention installing GMSimViz, only the dependencies. There is also no actual Python package for GMSimViz so it can't be installed using pip. See ucgmsim/GMSimViz#4 ucgmsim/GMSimViz#6
- [ ] Statement of need: There is a brief statement of need in the README but it could be clearer. The paper itself describes this a lot better so I'm not sure it needs to be in the README as well.
- [ ] Example usage: There is a short example in the README but it doesn't show the expected outcome. See ucgmsim/GMSimViz#7 and ucgmsim/GMSimViz#8
- [ ] Documentation: The docs are limited to a very brief description of the example data. See ucgmsim/GMSimViz#9
- [ ] Tests: There are no automated tests and the way the code is structured makes it very difficult to test. See ucgmsim/GMSimViz#10
- [ ] Community guidelines: There is no contributing guide or code of conduct. See ucgmsim/GMSimViz#11
- [ ] The GMT citation is outdated. See ucgmsim/GMSimViz#12

I short, the biggest issues that need to be addressed are regarding the lack of documentation and tests. The documentation can be added easily by the developers but testing will likely require refactoring of the code.

The resulting animations are very impressive! I had no idea that this could be done with GMT and it's very exciting to see someone pushing the boundaries of what the software can do. I think that with a bit of effort this will be a useful package for the geoscience community. The authors will not regret putting in the time to refactor the code for testing and packaging and creating top quality documentation. This is the kind of thing that takes a while to learn how to do but it pays off in the end because you have more confidence in your own code and avoid an lot of user emails asking simple questions :smile:
