+++ GMSimViz.py
@@ -16,11 +16,12 @@ from time import time, sleep
 
 from mpi4py import MPI
 import numpy as np
+
 try:
     from h5py import File as h5open
 except ImportError:
-    print('Missing h5py module.')
-    print('Will not be able to plot liquefaction or landslide.')
+    print("Missing h5py module.")
+    print("Will not be able to plot liquefaction or landslide.")
 
 import qcore.geo as geo
 import qcore.gmt as gmt
@@ -43,41 +44,59 @@ WINDOW_B = 0.3
 WINDOW_L = 0.5
 WINDOW_R = 0.5
 
+
 def load_xyts(meta):
     """
     Complete time-intensive tasks which aren't needed for beginning frames.
     """
-    xfile = xyts.XYTSFile(meta['xyts_file'], meta_only = False)
+    xfile = xyts.XYTSFile(meta["xyts_file"], meta_only=False)
     # higher res outline needed for mask to follow great circle line
     # not too high: becomes time intensive
-    geo.path_from_corners(corners = xfile.corners(), min_edge_points = 15, \
-            output = '%s/xyts/corners-hr.gmt' % (meta['wd']))
-    gmt.grd_mask('%s/xyts/corners-hr.gmt' % (meta['wd']), \
-            '%s/xyts/mask.nc' % (meta['wd']), \
-            dx = meta['xyts_res'], dy = meta['xyts_res'], \
-            region = meta['xyts_region'], wd = '%s/xyts' % (meta['wd']))
+    geo.path_from_corners(
+        corners=xfile.corners(),
+        min_edge_points=15,
+        output="%s/xyts/corners-hr.gmt" % (meta["wd"]),
+    )
+    gmt.grd_mask(
+        "%s/xyts/corners-hr.gmt" % (meta["wd"]),
+        "%s/xyts/mask.nc" % (meta["wd"]),
+        dx=meta["xyts_res"],
+        dy=meta["xyts_res"],
+        region=meta["xyts_region"],
+        wd="%s/xyts" % (meta["wd"]),
+    )
     # pgv used to generate cpt scale
-    xfile.pgv(pgvout = '%s/xyts/pgv.bin' % (meta['wd']))
-    cpt_max = gmt.xyv_cpt_range('%s/xyts/pgv.bin' % (meta['wd']))[2]
-    gmt.makecpt('magma', '%s/xyts/gm.cpt' % (meta['wd']), 0, cpt_max, \
-            invert = True, wd = '%s/xyts' % (meta['wd']), continuing = True)
-    xyts_cpt_max = gmt.xyv_cpt_range('%s/xyts/pgv.bin' % (meta['wd']))[2]
+    xfile.pgv(pgvout="%s/xyts/pgv.bin" % (meta["wd"]))
+    cpt_max = gmt.xyv_cpt_range("%s/xyts/pgv.bin" % (meta["wd"]))[2]
+    gmt.makecpt(
+        "magma",
+        "%s/xyts/gm.cpt" % (meta["wd"]),
+        0,
+        cpt_max,
+        invert=True,
+        wd="%s/xyts" % (meta["wd"]),
+        continuing=True,
+    )
+    xyts_cpt_max = gmt.xyv_cpt_range("%s/xyts/pgv.bin" % (meta["wd"]))[2]
 
     # prepare to use as overlay
-    if not os.path.isdir('%s/overlay' % (meta['wd'])):
+    if not os.path.isdir("%s/overlay" % (meta["wd"])):
         try:
-            os.makedirs('%s/overlay' % (meta['wd']))
+            os.makedirs("%s/overlay" % (meta["wd"]))
         except IOError:
             pass
[…]
