### 2018-09-10 1ec0682
and comparing to None fixed.
