+++ GMSimViz.py
@@ -277,7 +277,7 @@ def timeslice(job, meta):
         "%s_perspective%s.ps"
         % (
             os.path.splitext(os.path.basename(meta["srf_file"]))[0],
-            "_%.4d" % (i) * (job["seq"] != None),
+            "_%.4d" % (i) * (job["seq"] is not None),
         ),
     )
     if os.path.exists(
@@ -692,7 +692,7 @@ def timeslice(job, meta):
     elif plot == "timeseries":
         cpt_label = ""
         y = scale_p
-        if meta["xyts_file"] != None:
+        if meta["xyts_file"] is not None:
             x0 = WINDOW_L * 2.0
             x1 = scale_margin
             diff = x1 - x0
@@ -1006,7 +1006,7 @@ def get_args():
     except AssertionError:
         sys.exit("Could not find SRF: %s" % (args.srf_file))
     # xyts is optional
-    if args.xyts != None:
+    if args.xyts is not None:
         try:
             args.xyts = os.path.abspath(args.xyts)
             assert os.path.isfile(args.xyts)
@@ -1019,10 +1019,10 @@ def get_args():
         args.landslide_s,
         args.landslide_p,
     ]:
-        if filepath != None and not os.path.exists(filepath):
+        if filepath is not None and not os.path.exists(filepath):
             sys.exit("Could not find liquefaction/landslide file: %s" % (filepath))
     # paths optional
-    if args.paths != None:
+    if args.paths is not None:
         if not os.path.isdir(args.paths):
             sys.exit("Could not find path directory: %s" % (args.paths))
         path_sort = lambda n: map(
@@ -1080,7 +1080,7 @@ if len(sys.argv) > 1:
     )
 
     # working directory
-    if args.temp != None:
+    if args.temp is not None:
         gmt_temp = os.path.abspath(args.temp)
         print("Resuming from: %s" % (gmt_temp))
     else:
@@ -1167,7 +1167,7 @@ if len(sys.argv) > 1:
     frames_gm = 0
     final_azimuth = 180
     # xyts quick preparation
-    if xyts_file != None:
+    if xyts_file is not None:
         if not os.path.isdir(os.path.join(gmt_temp, "xyts")):
             os.makedirs(os.path.join(gmt_temp, "xyts"))
         xfile = xyts.XYTSFile(xyts_file, meta_only=True)
@@ -1185,7 +1185,7 @@ if len(sys.argv) > 1:
             msg_list.append([load_xyts, meta])
             msg_deps += 1
             frames_gm = int(xfile.dt * (xfile.nt - 0.6) * args.framerate)
-            if args.gm_cut != None:
+            if args.gm_cut is not None:
                 frames_gm = min(frames_gm, int(args.gm_cut * args.framerate))
         # ground motion 3D Z extent based on sim domain size
         # final size also depends on map tilt angle
@@ -1235,19 +1235,19 @@ if len(sys.argv) > 1:
     diff_azimuth /= frames_azimuth
 
     # prepare other data
-    if args.liquefaction_s != None:
+    if args.liquefaction_s is not None:
         region_liquefaction_s, poi_liquefaction_s = load_hdf5(
             args.liquefaction_s, "%s/overlay/liquefaction_s" % (gmt_temp)
         )
-    if args.liquefaction_p != None:
+    if args.liquefaction_p is not None:
         region_liquefaction_p, poi_liquefaction_p = load_hdf5(
             args.liquefaction_p, "%s/overlay/liquefaction_p" % (gmt_temp)
         )
-    if args.landslide_s != None:
+    if args.landslide_s is not None:
         region_landslide_s, poi_landslide_s = load_hdf5(
[…]
