+++ qcore/config.py
@@ -4,8 +4,7 @@ import sys
 
 import json
 
-config_file = os.path.join(os.path.dirname(os.path.abspath(__file__)), \
-        'config.json')
+config_file = os.path.join(os.path.dirname(os.path.abspath(__file__)), "config.json")
 
-with open(config_file, 'r') as f:
+with open(config_file, "r") as f:
     qconfig = json.load(f)
