+++ qcore/geo.py
@@ -8,8 +8,10 @@ import numpy as np
 
 R_EARTH = 6378.139
 
+
 class InputError(Exception):
     pass
+
 
 def gen_mat(mrot, mlon, mlat):
     """
@@ -30,16 +32,26 @@ def gen_mat(mrot, mlon, mlat):
     cosP = cos(arg)
     sinP = sin(arg)
 
-    amat = np.array([[cosA * cosT * cosP + sinA * sinP, \
-                      sinA * cosT * cosP - cosA * sinP, \
-                      sinT * cosP], \
-                     [cosA * cosT * sinP - sinA * cosP, \
-                      sinA * cosT * sinP + cosA * cosP, \
-                      sinT * sinP], \
-                     [-cosA * sinT, -sinA * sinT, cosT]], dtype = 'f')
+    amat = np.array(
+        [
+            [
+                cosA * cosT * cosP + sinA * sinP,
+                sinA * cosT * cosP - cosA * sinP,
+                sinT * cosP,
+            ],
+            [
+                cosA * cosT * sinP - sinA * cosP,
+                sinA * cosT * sinP + cosA * cosP,
+                sinT * sinP,
+            ],
+            [-cosA * sinT, -sinA * sinT, cosT],
+        ],
+        dtype="f",
+    )
     ainv = amat.T * 1.0 / np.linalg.det(amat)
 
     return amat.flatten(), ainv.flatten()
+
 
 def xy2ll(xy_km, amat):
     """
@@ -60,16 +72,17 @@ def xy2ll(xy_km, amat):
     yg = xp * amat[3] + yp * amat[4] + zp * amat[5]
     zg = xp * amat[6] + yp * amat[7] + zp * amat[8]
 
-    lat = np.where(zg == 0.0, 0.0, \
-                   90.0 - np.degrees(np.arctan(np.sqrt(xg * xg + yg * yg) / zg)))
+    lat = np.where(
+        zg == 0.0, 0.0, 90.0 - np.degrees(np.arctan(np.sqrt(xg * xg + yg * yg) / zg))
+    )
     lat[np.where(zg < 0.0)] -= 180.0
 
-    lon = np.where(xg == 0.0, 0.0, \
-                   np.degrees(np.arctan(yg / xg)))
+    lon = np.where(xg == 0.0, 0.0, np.degrees(np.arctan(yg / xg)))
     lon[np.where(xg < 0.0)] -= 180.0
     lon[np.where(lon < -180.0)] += 360.0
 
     return np.column_stack((lon, lat))
+
 
 def ll2xy(ll, ainv):
     """
@@ -89,8 +102,13 @@ def ll2xy(ll, ainv):
     zp = xg * ainv[6] + yg * ainv[7] + zg * ainv[8]
 
     # X km offsets from centre origin, Y km offsets from centre origin
-    return np.column_stack((R_EARTH * np.arcsin(yp / np.sqrt(1.0 - xp * xp)), \
-                            R_EARTH * np.arcsin(xp / np.sqrt(1.0 - yp * yp))))
+    return np.column_stack(
+        (
+            R_EARTH * np.arcsin(yp / np.sqrt(1.0 - xp * xp)),
+            R_EARTH * np.arcsin(xp / np.sqrt(1.0 - yp * yp)),
+        )
+    )
+
 
 def xy2gp(xy, nx, ny, hh):
     """
@@ -109,7 +127,8 @@ def xy2gp(xy, nx, ny, hh):
     # gridpoint from top corner
     gp /= hh
 
-    return np.round(gp).astype(np.int32, copy = False)
+    return np.round(gp).astype(np.int32, copy=False)
+
 
 def gp2xy(gp, nx, ny, hh):
     """
[…]
