### 2018-09-10 ucgmsim/GMSimViz#11 @viktor76525
I believe this issue is fixed. Comments, @leouieda ?
