### 2018-09-10 ucgmsim/GMSimViz#6 @viktor76525
@leouieda I'm not sure having all the scripts installable is possible. GMSimViz is an MPI program and therefore requires the main script to be run with multiple processes using something like openmpi mpirun or equivalent. The qcore functions include dependencies on the main script.
Another option is copying the functions currently under qcore into GMSimViz.py
Thoughts?
