### 2018-09-11 24f9de5
qcore becomes gmsimviz
