+++ gmsimviz/geo.py
@@ -280,201 +280,3 @@ def path_from_corners(
                 mp.write("%s %s\n" % (point[0], point[1]))
     else:
         return corners
-
-
-def wgs_nztm2000x(points):
-    """
-    Coordinates Transform: between WGS84 lon, lat and NZ Tranverse Mercator 2000
-    author: X. Bellagamba, Matlab -> Python by Viktor Polak
-
-    points: series of decimal longitude, latitude or nztm2k x, y
-    """
-    if type(points).__name__ == "list":
-        points = np.array(points)
-    points = np.atleast_2d(points)
-
-    if np.max(np.abs(points)) > 180:
-        wgs_out = True
-    else:
-        wgs_out = False
-        points = np.radians(points)
-
-    # NZTM2000 definitions (PROJ4 naming)
-    # origin
-    lon_0 = radians(173.0)
-    lat_0 = 0.0
-    # false northing, easting (metres)
-    y_0 = 10000000
-    x_0 = 1600000
-    # central meridian scaling factor
-    k_0 = 0.9996
-    # flattening of the ellipsoid (%)
-    f = 1 / 298.257222101
-    # semimajor, semiminor ellipsoid radius
-    a = 6378137.0
-    b = a * (1 - f)
-    # eccentricity of the ellipsoid squared
-    es = (f + f) - (f * f)
-
-    # conversion specific
-    A_0 = 1 - es / 4.0 - 3 * es ** 2 / 64.0 - 5 * es ** 3 / 256.0
-    A_2 = 3 / 8.0 * (es + es ** 2 / 4.0 + 15 * es ** 3 / 128.0)
-    A_4 = 15 / 256.0 * (es ** 2 + 3 * es ** 3 / 4.0)
-    A_6 = 35 * es ** 3 / 3072.0
-    m_0 = a * (
-        A_0 * lat_0 - A_2 * sin(2 * lat_0) + A_4 * sin(4 * lat_0) - A_6 * sin(6 * lat_0)
-    )
-    n = (a - b) / (a + b)
-    G = a * (1 - n) * (1 - n ** 2) * radians(1 + 9 * n ** 2 / 4.0 + 225 * n ** 4 / 64.0)
-
-    ###
-    ### process all points
-    ###
-    if wgs_out:
-        N_prime = points[:, 1] - y_0
-        E_prime = points[:, 0] - x_0
-        m_prime = m_0 + N_prime / k_0
-        sigma = np.radians(m_prime) * 1 / G
-        phi_prime = (
-            sigma
-            + (3 * n / 2.0 - 27 * n ** 3 / 32.0) * np.sin(2 * sigma)
-            + (21 * n ** 2 / 16.0 - 55 * n ** 4 / 32.0) * np.sin(4 * sigma)
-            + (151 * n ** 3 / 96.0) * np.sin(6 * sigma)
-            + (1097 * n ** 4 / 512.0) * np.sin(8 * sigma)
-        )
-        y_factors = phi_prime
-    else:
-        y_factors = points[:, 1]
-        m = a * (
-            A_0 * y_factors
-            - A_2 * np.sin(2 * y_factors)
-            + A_4 * np.sin(4 * y_factors)
-            - A_6 * np.sin(6 * y_factors)
-        )
-        omega = points[:, 0] - lon_0
-
-    # ellipsoid radius in the prime vertical
-    nu = a / np.sqrt(1 - es * np.sin(y_factors) ** 2)
-    # projection parameters
-    rho = (a * (1 - es)) / ((1 - es * np.sin(y_factors) ** 2) ** 1.5)
-    psi = nu / rho
-    t = np.tan(y_factors)
-    if wgs_out:
-        rs = rho * nu * k_0 ** 2
[…]
