### 2018-09-11 ucgmsim/GMSimViz#4 @viktor76525
qcore -> gmsimviz
setup.py installs the main script into $PATH
