### 2018-09-11 ucgmsim/GMSimViz#11 @viktor76525
Please open another issue if there are any further problems.
