### 2018-09-11 b6519c9
minor readme changes, more to come
