### 2018-09-11 e15a8f0
fixed logo path
