### 2018-09-14 109acb7
added smaller examples and wrote out details of all parameters
