+++ README.md
@@ -79,29 +79,49 @@ Currently unsupported (we cannot give instructions but if the dependencies are m
 
 
 ## Sample Data
-Sample data is provided, it contains smaller data sets from multiple earthquakes to make a feature-full demo. To produce the sample animation with 8 processes, run the following from the repository location upon installation:
+Sample data is provided, it contains smaller data sets from multiple earthquakes to make a feature-full demo. 
+
+The most simplest output is a single frame (image) facing the fault plane. This case only requires the SRF file as the parameter:
+<img src="figures/fault_perspective.jpg" width="400">
+```shell
+gmsimviz sample_data/fault.srf
+```
+
+An animation can be created from an SRF that shows the rupture propagating. It can be expected to take around 12 minutes to complete on a decent personal machine. Here is an example using 7 slave processes for a machine with 8 cores (including hyperthreading). An example of what is expected is available at https://youtu.be/nTtq_DxGVUM
+```shell
+gmsimviz sample_data/fault.srf -a --crude -n7 --title "Fault Animation" --dpi 120 --downscale 1
+```
+
+The fully featured sample animation below takes about 45 minutes with -n7 on a v4 E5-2620 (2.1 GHz). Slave processes are spawned after about 3 minutes and frames will start appearing in the temporary folder (in the same folder as the SRF file, sample_data/_GMT_WD_PERSPECTIVE_<random sequence>). When complete, the video is available in the working directory and the temporary folder is removed by default.
 ```shell
 gmsimviz sample_data/fault.srf -a --crude -n7 --title "Sample Animation" --dpi 120 --downscale 1 -x sample_data/xyts.e3d --liquefaction-s sample_data/liquefaction_s.hdf5 --liquefaction-p sample_data/liquefaction_p.hdf5 --landslide-s sample_data/landslide_s.hdf5 --landslide-p sample_data/landslide_p.hdf5 --paths sample_data/transport
 ```
-If you don't have `python2` in your `env`, you may need to run with `python2 GMSimViz.py` or `python GMSimViz.py` instead of `./GMSimViz.py`.\
-Most parameters in the above example are optional. Some provide more segments in the animation, others have defaults. Here is a summary of the sample command options:
+
+Here are details for command options:
 
 - `srf_file` first position, not optional. Path to a fault in standard rupture format containing the plane header (https://scec.usc.edu/scecpedia/Standard_Rupture_Format).
 - `-a` create animation instead of static image showing the slip distribution only. Will animate the progression of slip if no other input data are given.
-- `--crude` this will produce a quick result. No topography or roads are plotted, low resolution coastlines are used and overlays have lower resolutions.
+- `--crude` this will produce a quick result. No topography or roads are plotted, low resolution coastlines are used and overlays have lower resolutions. This is required for faults outside of the New Zealand area because the higher resolution topography is only available here. Alternatively, replace the topography files with your own.
 - `-n28` number of slave (worker) processes to start. 7 is a good choice for systems with only 8 cores. It is recommended to run on a machine with many cores.
-- `-f10` set the framerate.
+- `-f10` set the framerate of the animation.
 - `--title` title on the movie, the default is the basename of the SRF file.
-- `--dpi` dpi of output. Frames are 16 inches x 9 inches so 240 will produce 4k output.
+- `--dpi` dpi of output. Frames are 16 inches x 9 inches so 240 will produce 4k output, 120 for FullHD.
 - `--downscale` render at higher resolution and then downscale. Prevents jitter in object positioning. 8 is ideal for a smooth result.
 - `-x` XYTS file. This is an output created by the EMOD3D software that simulates ground motion. This provides the simulation domain and ground motion data.
[…]
