### 2018-09-14 ucgmsim/GMSimViz#8 @viktor76525
two more simple examples with sample outputs added. one takes about 2 minutes, the other 12.
