### 2018-09-14 ucgmsim/GMSimViz#7 @viktor76525
All 3 examples have expected outputs available.
