### 2018-09-16 #808 @viktor76525
Minor changes fixed, samples to run include quicker options (2 minutes to 12 minutes instead of 45 minutes). The output for these and the original are available to compare.
Will expand the documentation a bit.
Can break long functions into smaller ones to an extent.
Do not intend to create unit tests, will go with the option to give sample/test data instead.
