### 2018-09-16 f51c632
added git install instructions
