### 2018-09-16 ucgmsim/GMSimViz#6 @viktor76525
expanded instructions for installing from git.
The main issue with packaging is that GMSimViz requires large geographic datasets for full functionality.
Alternatively the datasets could be hardcoded to the users home directory or system and be made optional.
