### 2018-09-17 24c3754
more descriptive command line options
