### 2018-09-17 ucgmsim/GMSimViz#9 @viktor76525
much more detail provided
