### 2018-10-02 #808 @viktor76525
@lheagy and @leouieda I have closed all the issues, ready for checking over.
