### 2018-10-02 ucgmsim/GMSimViz#10 @viktor76525
much shorter functions
