### 2018-11-01 #808 @lheagy
👋 Hi @leouieda and @hugoledoux, thanks for your first round of reviews. Would you have time to take a look through @viktor76525's changes?  Thanks!
