### 2018-12-06 #808 @hugoledoux
ok @viktor76525 I tried with a clean install under macOS and Python 3.6.5.

1. I can install all Python dependencies successfully.
2. GMT installation is described with MacPort and Fink, but both are old IMO, and there is no way I install such beast in my system! Better to use [homebrew](https://brew.sh/), and it's there and it works.
3. once all is setup and I have downloaded the .zip, I get errors but I have no idea what they mean:

```
(gsim2) wlan-145-94-215-233:GMSimViz hugo$ python setup.py install --user
running install
Checking .pth file support in /Users/hugo/Library/Python/3.6/lib/python/site-packages/
/Users/hugo/.pyenv/versions/3.6.5/envs/gsim2/bin/python -E -c pass
TEST FAILED: /Users/hugo/Library/Python/3.6/lib/python/site-packages/ does NOT support .pth files
error: bad install directory or PYTHONPATH

You are attempting to install a package to a directory that is not
on PYTHONPATH and which Python does not read ".pth" files from.  The
installation directory you specified (via --install-dir, --prefix, or
the distutils default setting) was:

    /Users/hugo/Library/Python/3.6/lib/python/site-packages/

and your PYTHONPATH environment variable currently contains:

    ''

Here are some of your options for correcting the problem:

* You can choose a different installation directory, i.e., one that is
  on PYTHONPATH or supports .pth files

* You can add the installation directory to the PYTHONPATH environment
  variable.  (It must then also be on PYTHONPATH whenever you run
  Python and want to use the package(s) you are installing.)

* You can set up the installation directory to support ".pth" files by
  using one of the approaches described here:

  https://setuptools.readthedocs.io/en/latest/easy_install.html#custom-installation-locations

Please make the appropriate changes for your system and try again.
```

Do you? What is a .pth file?
