### 2018-12-06 #808 @hugoledoux
ok I realised that there's no xrange() function in Python3... I guess the port to Python3 has not been made for everything?
