+++ gmsimviz/gmt.py
@@ -109,7 +109,7 @@ def update_gmt_path(gmt_bin, wd=None):
     GMT = gmt_bin
     # retrieve version of GMT
     gmtp = Popen([GMT, "--version"], stdout=PIPE)
-    GMT_VERSION = gmtp.communicate()[0].rstrip().decode()
+    GMT_VERSION = gmtp.communicate()[0].rstrip().decode("utf-8")
     GMT_MAJOR, GMT_MINOR = map(int, GMT_VERSION.split(".")[:2])
 
     psconvert = "psconvert"
@@ -261,7 +261,7 @@ def simplify_segs(infile, outfile=None):
 
     # return as list of (individual) points
     proc = Popen(["awk", segfile_simple_awk, infile], stdout=PIPE)
-    result = proc.communicate()[0]
+    result = proc.communicate()[0].decode("utf-8")
     proc.wait()
     return np.loadtxt(result.split("\n"), dtype="f")
 
@@ -400,7 +400,7 @@ def make_seismo(
         )
 
     elif fmt == "time":
-        for t in xrange(len(tsy)):
+        for t in range(len(tsy)):
             tsyp = np.copy(tsy[t::-1]) * yfac + y0 - yfac * tsy[t]
             tsx = np.arange(len(tsyp)) * xfac + x0
             np.savetxt(
@@ -430,7 +430,7 @@ def auto_tick(x_min, x_max, width):
 
     # starting tick is increased until ticks per inch is less than max
     major_tick = 0.01
-    for i in xrange(12):
+    for i in range(12):
         # check tpi vs tpi max for decimal places in major_tick
         if ((x_max - x_min) / major_tick) / width > tpi_dp[max(0, 2 - i // 3)]:
             # increase by factor of 2, 2.5, 2, 2, 2.5, 2, 2, 2.5...
@@ -535,7 +535,7 @@ def is_native_xyv(xyv_file, x_min, x_max, y_min, y_max, v_min=None):
     bin_data = np.fromfile(xyv_file, dtype="3f4")
 
     # check the first few rows
-    for i in xrange(min(10, len(bin_data))):
+    for i in range(min(10, len(bin_data))):
         if (
             x_min <= bin_data[i, 0] <= x_max
             and y_min <= bin_data[i, 1] <= y_max
@@ -717,7 +717,7 @@ def srf2map(
         )
 
     # create resources for each plane
-    for s in xrange(n_plane):
+    for s in range(n_plane):
         if not xy:
             # geographical based resources
             x_min, y_min = np.min(np_bounds[s], axis=0)
@@ -1106,7 +1106,7 @@ def table2grd(
         try:
             # test if text file
             with open(table_in, "r") as tf:
-                for _ in xrange(header):
+                for _ in range(header):
                     tf.readline()
                 # assert added to catch eg: first line = '\n'
                 assert len(map(float, tf.readline().split()[:2])) == 2
@@ -1114,14 +1114,14 @@ def table2grd(
             cmd.append("-bi3f")
         # run command
         p = Popen(cmd, stderr=PIPE, cwd=wd)
-        e = p.communicate()[1]
+        e = p.communicate()[1].decode("utf-8")
         p.wait()
         # also create radius based mask if wanted
         if automask != None:
             Popen(cmd_mask, cwd=wd).wait()
     else:
         p = Popen(cmd, stdin=PIPE, stderr=PIPE, cwd=wd)
-        e = p.communicate(table_in)[1]
+        e = p.communicate(table_in.encode("utf-8"))[1].decode("utf-8")
         p.wait()
 
     write_history(True, wd=wd)
@@ -1175,7 +1175,7 @@ def grdclip(
         cmd.append("-R%s/%s" % ("/".join(map(str, region)), new))
[…]
