### 2018-12-07 #808 @viktor76525
Thanks for pointing that out @hugoledoux . Last time Python 3 was meant to have been tested, Python 2 must have managed to sneak in because there is no way it would have worked.
All should be good with Python 3 now, I have tested the first and last samples with Python 3.6.5.
