### 2018-12-17 #808 @hugoledoux
sorry for delay again, the python3 error is gone but no visualisation:
```
$ ./gmsimviz ../sample_data/fault.srf
grdimage: Error: (x_max-x_min) must equal (NX + eps) * x_inc), where NX is an integer and |eps| <= 0.0001.
grdimage (gmtlib_read_grd_info): Use grdedit -A on your grid file to make region and increments compatible [/Users/hugo/temp/GMSimViz/resources/Topo/srtm_all_filt_nz_i5.grd]
grdimage: Error: (x_max-x_min) must equal (NX + eps) * x_inc), where NX is an integer and |eps| <= 0.0001.
grdimage (gmtlib_read_grd_info): Use grdedit -A on your grid file to make region and increments compatible [/Users/hugo/temp/GMSimViz/resources/Topo/srtm_all_filt_nz.grd]
```
