### 2018-12-17 #808 @viktor76525
Hi @hugoledoux, that is normal as of ~5.4.4 and I think it doesn't classify as an error in 5.4.3.

That particular command (first example) only produces a static image (png). You should have a file `fault_perspective.png` where you run the command from. If it has not created the image, were there any other messages?
