### 2018-12-18 #808 @hugoledoux
oh okay it did work then. Perhaps writing this in the doc for newbies like me if a good idea though? I saw the ERROR and thought things weren't working...

I'll finish the review soon then.
