### 2018-12-18 #808 @hugoledoux
Some things that need to be fixed/modified:

# installation

- GMT: I'd add for mac that homebrew can be used, as I wrote above.
- git is not necessary to download the software, don't put it as a dependency

# running software

maybe a bit more clarity about what the output is and that there'll be error

#
