### 2018-12-19 ede6ee0
git is not a dependency, only for compiling GMT
