### 2018-12-19 5972099
ignore pre-computed grids error in new GMT
