### 2018-12-19 #808 @viktor76525
I've made all the changes.
- message about only a single png created in first example readme: "The most simplest output is a single frame (image) facing the fault plane. ..." not sure if this is clear enough.
- git no longer listed as it is part of the GMT build process (optional).
- homebrew listed as Mac option
