### 2019-01-07 #808 @lheagy
Thanks @hugoledoux!

@viktor76525: testing is a mandatory component for acceptance. Ideally, these are automated tests, but is also acceptable to have documentation outlining steps that the user can take to test and verify that the software is functioning as expected. @leouieda provided some suggestions in ucgmsim/GMSimViz#10 for setting up automated tests. Many of the functions in [geo.py](https://github.com/ucgmsim/GMSimViz/blob/master/gmsimviz/geo.py) are well-suited for setting up some simple tests as suggested.

Even if you do not unit-test the software as suggested in that issue, you could still implement tests that test the output of larger workflows.

@viktor76525: In order to give us an idea of how you plan to proceed, would you mind opening a new issue in your repo or re-opening ucgmsim/GMSimViz#10 and sketching out how you plan to proceed with testing of the software?
