### 2019-01-07 #808 @viktor76525
Hi @lheagy.
I was hoping to go for the "OK" section at https://joss.readthedocs.io/en/latest/review_criteria.html

> Good: An automated test suite hooked up to an external service such as Travis-CI or similar
> OK: Documented manual steps that can be followed to check the expected functionality of the software (e.g. a sample input file to assert behaviour)
> Bad (not acceptable): No way for you the reviewer to check whether the software works

I believe the [GMSimViz](https://github.com/ucgmsim/GMSimViz) page does contain 3 such manual steps (sample commands) under the Sample Data section along with expected outputs.
