### 2019-01-07 #808 @lheagy
Okay, I will bring this to the editorial team for input.
