### 2019-01-09 #808 @lheagy
Hi @viktor76525, I brought this up with the editorial board and the consensus is that at a minimum, we require that there be an **objective** way to test the code. Running an example and eye-ball comparing the output to a video is a subjective test.

The reason that we require at least some objective testing is so that if you or a contributor makes a change to the code, how do you know it didn't break? Since you refactored the code into smaller functions, some of the core functionality should be much easier to design tests for. You do not need to test the entire code-base to meet JOSS criteria, but at a minimum, there should be a mechanism for someone to assess if core functionality is working as expected.

Please let me know if I can provide further clarification.
