### 2019-01-17 #808 @leouieda
@viktor76525 I've been thinking about ways that you can check that the end result of the software remains the same when you change the code (a regression test). This is not as good as testing the individual functionality of all pieces of the code but it can at least give you the piece of mind of knowing that your changes didn't break the existing functionality.

One way you could do this is to:
1. Store a reference video or set of images that you judge to be correct
2. Use the `compare` command from imagemagick to calculate the difference between the files you stored and new output from the software.
3. Check if the difference is below a threshold.

This wouldn't require any testing framework or much extra coding. A shell script that run the program and compares all images should be relatively straightforward. Let me know if you want help implementing any of this.

@lheagy would this be enough to satisfy the JOSS requirements?
