### 2019-01-18 #808 @viktor76525
Thanks @leouieda, I'll have a look into it. So far I've made pytest tests for the smaller library functions.
