+++ tests/gmt_test_mpi.py
@@ -1,255 +1,0 @@
-#!/usr/bin/env python2
-"""
-MPI pattern based on
-github.com/jbornschein/mpi4py-examples/blob/master/09-task-pull.py
-"""
-
-from hashlib import sha1
-import os
-from shutil import rmtree
-import sys
-from time import time
-
-from scipy.misc import imread
-from mpi4py import MPI
-
-# first test is loading library
-from gmsimviz import gmt
-
-# MPI stuff
-comm = MPI.COMM_WORLD
-rank = comm.Get_rank()
-size = comm.Get_size()
-if size == 1:
-    sys.exit("Use mpirun/similar with more than 1 process.")
-status = MPI.Status()
-MASTER = 0
-class tags:
-    READY = 15
-    DONE = 85
-    START = 204
-    EXIT = 240
-
-# GMT versions to test
-GMT_PATHS = { \
-    '5.1.2':'/opt/gmt-5.1.2/bin/gmt', \
-    '5.3.2':'/opt/gmt-5.3.2/bin/gmt', \
-    '5.3.3':'/opt/gmt-5.3.3/bin/gmt', \
-    '5.4.1':'/opt/gmt-5.4.1/bin/gmt', \
-    '5.4.2':'/opt/gmt-5.4.2/bin/gmt', \
-    '5.4.3':'/opt/gmt-5.4.3/bin/gmt', \
-    '5.4.4':'/opt/gmt-5.4.4/bin/gmt'}
-
-test_dir = os.path.abspath('GMT_TESTING')
-if rank == MASTER:
-    if os.path.exists(test_dir):
-        rmtree(test_dir)
-    os.makedirs(test_dir)
-
-###
-### TESTING FUNCTIONS
-###
-
-def test_coastlines(pf):
-    p = gmt.GMTPlot(pf)
-    p.spacial('M', (170, 175, -44, -38), sizing = 3)
-    p.coastlines(width = '0.4p')
-    p.finalise()
-    p.png(dpi = 200, clip = True)
-
-def test_land(pf):
-    p = gmt.GMTPlot(pf)
-    p.spacial('M', (170.1, 179.91, -37, -34), sizing = 7)
-    p.land(fill = 'darkred')
-    p.finalise()
-    p.png(dpi = 100, clip = True)
-
-def test_ticks(pf):
-    p = gmt.GMTPlot(pf)
-    p.spacial('M', (160.992, 174.9122, -44, -34.01), sizing = 2)
-    p.ticks(major = '1d', minor = '20m', sides = 'ew')
-    p.finalise()
-    p.png(dpi = 100, clip = True)
-
-def test_ticks2(pf):
-    p = gmt.GMTPlot(pf)
-    p.spacial('T', (160.992, 174.9122, -44, -34.01), sizing = 5, \
-            x_shift = 0.5, y_shift = 0.5, lon0 = 174.9122)
-    p.ticks(major = '1d', minor = '20m', sides = 'ew')
-    p.finalise()
-    p.png(dpi = 100, clip = True)
-
-def test_cpt(pf):
-    cptf = '%s.cpt' % (os.path.splitext(pf)[0])
-    gmt.makecpt('hot', cptf, 0, 120, inc = 0.1, invert = True, \
-            wd = os.path.dirname(pf))
-    p = gmt.GMTPlot(pf)
-    p.spacial('X', (0, 15, 0, 4), sizing = '15/4')
-    p.cpt_scale(6.1, 2.05, cptf, 20, 5, label = 'test_scale', \
-            length = 3.05, thickness = '0.3i')
-    p.finalise()
-    p.png(dpi = 320, clip = True)
-
-def test_cpt2(pf):
[…]
