### 2019-01-18 #808 @lheagy
Great suggestion @leouieda! Yes, this would satisfy the JOSS requirements.
