### 2019-01-18 #808 @leouieda
> So far I've made pytest tests for the smaller library functions.

That's great! Don't hesitate to ask if you need any feedback or help with that too. The regression test can also be useful even if you have unit tests as well.
