+++ setup.py
@@ -13,7 +13,7 @@ with open(os.path.join(script_dir, "gmsimviz", "config.json"), "w") as cfg:
 
 setup(
     name="GMSimViz",
-    version="1.2",
+    version="1.2.1",
     packages=["gmsimviz"],
     url="https://github.com/ucgmsim/GMSimViz",
     description="Ground Motion Visualisation",
