### 2019-01-27 #808 @lheagy
Excellent @viktor76525! @leouieda, would you mind taking a look at the tests that have been added?
