### 2019-02-04 #808 @lheagy
:wave: Hi @leouieda and @hugoledoux, would either of you have time to take a look at the testing additions this week? If not, please let me know and I can go through. Thanks!
