### 2019-02-06 #808 @leouieda
@lheagy I'm still on the road so it's a bit tough to try to run the tests. Sorry for the delay.
