### 2019-02-10 ucgmsim/GMSimViz#19 @lheagy
test expecting a config file
It looks like the test `test_gmt` requires a config file. Is there one that can be included in the test directory?
```
______________________ ERROR collecting tests/test_gmt.py ______________________
test_gmt.py:14: in <module>
    from gmsimviz import gmt
../../../gmsimviz/gmt.py:22: in <module>
    from gmsimviz.config import config
../../../gmsimviz/config.py:8: in <module>
    with open(config_file, "r") as f:
E   FileNotFoundError: [Errno 2] No such file or directory: '/.../gmsimviz/config.json'
```

related to the JOSS review openjournals/joss-reviews#808
