### 2019-02-10 #808 @lheagy
Hi @viktor76525, the tests are looking good. I just ran into the error noted here: ucgmsim/GMSimViz#19. Would you mind taking a look?
