### 2019-02-10 ucgmsim/GMSimViz#19 @viktor76525
_config.json_ is generated as part of running `setup.py install` which is part of the install instructions.

i have added a default to the repository to make pytest run even without installing gmsimviz

at least 2 tests will still fail unless the plotting resources are downloaded as per instructions on main github page or when `setup.py install` attempted
