### 2019-02-10 ucgmsim/GMSimViz#19 @viktor76525
download archive here:
https://goo.gl/mYFCQn

exctract such that the directory exists:
`<GMSimViz repo dir>/resources`
