### 2019-03-05 ucgmsim/GMSimViz#20 @leouieda
Missing Scipy dependency
The README doesn't specify it but it's included in `setup.py`. Is scipy needed?

openjournals/joss-reviews#808
