### 2019-03-05 ucgmsim/GMSimViz#20 @leouieda
Trying to run the tests I see that scipy is needed. But version >=1.2.0 won't work because `scipy.misc.imread` is no longer in Scipy (https://docs.scipy.org/doc/scipy/reference/generated/scipy.misc.imread.html).
