### 2019-03-05 ucgmsim/GMSimViz#20 @leouieda
Furthermore, even with scipy 1.1.0 the `pillow` library is required to access `imread` and it's not installed with scipy by default (it must be installed by the user). So it would be great to have it listed as a dependency here as well (if only an indirect one).
