### 2019-03-05 ucgmsim/GMSimViz#21 @leouieda
Tests fail with GMT 5.4.5
I managed to run the tests with `pytest` using GMT 5.4.5 and got several failures from `tests/test_gmt.py`  (see below). Comparing the hashes of the expected PNGs will not work reliably between machines, OSs, GMT versions, ghostscript versions, etc, because some likes will be slightly different or even the amount of whitespace might be different. All of these would cause the hash to change when the image itself didn't change significantly. The hash is also doesn't provide any information about what changed and if this change is significant.

The method used by matplotlib and GMT itself is to calculate a different between the PNGs and compare them with a threshold. This way, you can also see the difference to judge is it's just the font being slightly different or if the plot is actually wrong.

With a bit of hacking to make a `Figure` class with a `savefig` method, you could get [pytest-mpl](https://github.com/matplotlib/pytest-mpl) to run all of these checks for you. That module requires you to add a decorator to the test function and make it return an object that has a `savefig` method. The method needs to take a file name and generate a PNG with that name. In your case, this is what your test code already does so it would be a matter of making the class and moving the code you have into it.

openjournals/joss-reviews#808

<summary>full error log from pytest (click to expand)</summary>
<pre>
_____________________________________________________________________ test_all[__test_coastlines-61efdfbe4cd9bfd5a90ad86c67013c8d9494abc6-5.0] _____________________________________________________________________

test = <function __test_coastlines at 0x7fcaaa0a0488>, expected_hash = '61efdfbe4cd9bfd5a90ad86c67013c8d9494abc6', min_version = 5.0

    @pytest.mark.parametrize("test, expected_hash, min_version", [
        (__test_coastlines, '61efdfbe4cd9bfd5a90ad86c67013c8d9494abc6', 5.0),
        (__test_land, '8c31ef6345aaad068cc4eb3e1d4b3f78cd6fc9a3', 5.0),
        (__test_ticks, 'd625bfb10464664d38f60537686999945eafafe7', 5.0),
        (__test_ticks2, '903830d2ba5d19a415a6d4e15fa62175bd7b7242', 5.0),
        (__test_cpt, '960032a10ef5fef8f013aab8892082fc4a606c9c', 5.2),
        (__test_cpt2, 'd83a3f811e8e514a6e88dacf38fe24bc4dea3280', 5.2),
        (__test_fill, 'eaa069d015eefb20f9826061a6de09d17a420a91', 5.0),
        (__test_autotick, None, 5.0),
        (__test_autotick2, None, 5.0),
        (__test_autotick3, None, 5.0),
    ])
    def test_all(test, expected_hash, min_version):
        wd = os.path.join(TEMP_DIR, test.__name__)
        if not os.path.exists(wd):
            os.makedirs(wd)

        if float('.'.join(gmt.GMT_VERSION.split('.')[:2])) < min_version:
            return

        if expected_hash != None:
            ps = '%s/%s.ps' % (wd, test.__name__)
            png = '%s/%s.png' % (wd, test.__name__)
            test(ps)
            os.symlink(png, os.path.join(TEMP_DIR, os.path.basename(png)))
            assert os.path.isfile(png)
            png_hash = sha1(imread(png)).hexdigest()
>           assert png_hash == expected_hash
E           AssertionError: assert '043368365713...c73d3d23c4c57' == '61efdfbe4cd9b...13c8d9494abc6'
E             - 04336836571352041fc1a2dda7cc73d3d23c4c57
E             + 61efdfbe4cd9bfd5a90ad86c67013c8d9494abc6

tests/test_gmt.py:136: AssertionError
----------------------------------------------------------------------------------------------- Captured stderr call -----------------------------------------------------------------------------------------------
psxy: Error for input file: No such file (/tmp/pip-req-build-vicroigy/resources/Paths/lds-nz-coastlines-and-islands/150k.gmt)
psconvert: Unable to decode BoundingBox file ./psconvert_15752c.bb
psconvert: Failed to remove ./psconvert_15752c.bb! [remove error: No such file or directory]
________________________________________________________________________ test_all[__test_land-8c31ef6345aaad068cc4eb3e1d4b3f78cd6fc9a3-5.0] ________________________________________________________________________

test = <function __test_land at 0x7fca94be9d90>, expected_hash = '8c31ef6345aaad068cc4eb3e1d4b3f78cd6fc9a3', min_version = 5.0

    @pytest.mark.parametrize("test, expected_hash, min_version", [
        (__test_coastlines, '61efdfbe4cd9bfd5a90ad86c67013c8d9494abc6', 5.0),
        (__test_land, '8c31ef6345aaad068cc4eb3e1d4b3f78cd6fc9a3', 5.0),
        (__test_ticks, 'd625bfb10464664d38f60537686999945eafafe7', 5.0),
        (__test_ticks2, '903830d2ba5d19a415a6d4e15fa62175bd7b7242', 5.0),
        (__test_cpt, '960032a10ef5fef8f013aab8892082fc4a606c9c', 5.2),
        (__test_cpt2, 'd83a3f811e8e514a6e88dacf38fe24bc4dea3280', 5.2),
        (__test_fill, 'eaa069d015eefb20f9826061a6de09d17a420a91', 5.0),
        (__test_autotick, None, 5.0),
        (__test_autotick2, None, 5.0),
        (__test_autotick3, None, 5.0),
    ])
    def test_all(test, expected_hash, min_version):
        wd = os.path.join(TEMP_DIR, test.__name__)
        if not os.path.exists(wd):
            os.makedirs(wd)

        if float('.'.join(gmt.GMT_VERSION.split('.')[:2])) < min_version:
            return

        if expected_hash != None:
            ps = '%s/%s.ps' % (wd, test.__name__)
            png = '%s/%s.png' % (wd, test.__name__)
            test(ps)
            os.symlink(png, os.path.join(TEMP_DIR, os.path.basename(png)))
            assert os.path.isfile(png)
            png_hash = sha1(imread(png)).hexdigest()
>           assert png_hash == expected_hash
E           AssertionError: assert '043368365713...c73d3d23c4c57' == '8c31ef6345aaa...b3f78cd6fc9a3'
E             - 04336836571352041fc1a2dda7cc73d3d23c4c57
E             + 8c31ef6345aaad068cc4eb3e1d4b3f78cd6fc9a3

tests/test_gmt.py:136: AssertionError
----------------------------------------------------------------------------------------------- Captured stderr call -----------------------------------------------------------------------------------------------
psxy: Error for input file: No such file (/tmp/pip-req-build-vicroigy/resources/Paths/lds-nz-coastlines-and-islands/150k.gmt)
psconvert: Unable to decode BoundingBox file ./psconvert_15796c.bb
psconvert: Failed to remove ./psconvert_15796c.bb! [remove error: No such file or directory]
_______________________________________________________________________ test_all[__test_ticks-d625bfb10464664d38f60537686999945eafafe7-5.0] ________________________________________________________________________

test = <function __test_ticks at 0x7fca94beb158>, expected_hash = 'd625bfb10464664d38f60537686999945eafafe7', min_version = 5.0

    @pytest.mark.parametrize("test, expected_hash, min_version", [
        (__test_coastlines, '61efdfbe4cd9bfd5a90ad86c67013c8d9494abc6', 5.0),
        (__test_land, '8c31ef6345aaad068cc4eb3e1d4b3f78cd6fc9a3', 5.0),
        (__test_ticks, 'd625bfb10464664d38f60537686999945eafafe7', 5.0),
        (__test_ticks2, '903830d2ba5d19a415a6d4e15fa62175bd7b7242', 5.0),
        (__test_cpt, '960032a10ef5fef8f013aab8892082fc4a606c9c', 5.2),
        (__test_cpt2, 'd83a3f811e8e514a6e88dacf38fe24bc4dea3280', 5.2),
        (__test_fill, 'eaa069d015eefb20f9826061a6de09d17a420a91', 5.0),
        (__test_autotick, None, 5.0),
        (__test_autotick2, None, 5.0),
        (__test_autotick3, None, 5.0),
    ])
    def test_all(test, expected_hash, min_version):
        wd = os.path.join(TEMP_DIR, test.__name__)
        if not os.path.exists(wd):
            os.makedirs(wd)

        if float('.'.join(gmt.GMT_VERSION.split('.')[:2])) < min_version:
            return

        if expected_hash != None:
            ps = '%s/%s.ps' % (wd, test.__name__)
            png = '%s/%s.png' % (wd, test.__name__)
            test(ps)
            os.symlink(png, os.path.join(TEMP_DIR, os.path.basename(png)))
            assert os.path.isfile(png)
            png_hash = sha1(imread(png)).hexdigest()
>           assert png_hash == expected_hash
E           AssertionError: assert 'd5a8d0066853...ee51ede9c7cb2' == 'd625bfb104646...999945eafafe7'
E             - d5a8d0066853263f9350303833aee51ede9c7cb2
E             + d625bfb10464664d38f60537686999945eafafe7

tests/test_gmt.py:136: AssertionError
_______________________________________________________________________ test_all[__test_ticks2-903830d2ba5d19a415a6d4e15fa62175bd7b7242-5.0] _______________________________________________________________________

test = <function __test_ticks2 at 0x7fca94beb1e0>, expected_hash = '903830d2ba5d19a415a6d4e15fa62175bd7b7242', min_version = 5.0

    @pytest.mark.parametrize("test, expected_hash, min_version", [
        (__test_coastlines, '61efdfbe4cd9bfd5a90ad86c67013c8d9494abc6', 5.0),
        (__test_land, '8c31ef6345aaad068cc4eb3e1d4b3f78cd6fc9a3', 5.0),
        (__test_ticks, 'd625bfb10464664d38f60537686999945eafafe7', 5.0),
        (__test_ticks2, '903830d2ba5d19a415a6d4e15fa62175bd7b7242', 5.0),
        (__test_cpt, '960032a10ef5fef8f013aab8892082fc4a606c9c', 5.2),
        (__test_cpt2, 'd83a3f811e8e514a6e88dacf38fe24bc4dea3280', 5.2),
        (__test_fill, 'eaa069d015eefb20f9826061a6de09d17a420a91', 5.0),
        (__test_autotick, None, 5.0),
        (__test_autotick2, None, 5.0),
        (__test_autotick3, None, 5.0),
    ])
    def test_all(test, expected_hash, min_version):
        wd = os.path.join(TEMP_DIR, test.__name__)
        if not os.path.exists(wd):
            os.makedirs(wd)

        if float('.'.join(gmt.GMT_VERSION.split('.')[:2])) < min_version:
            return

        if expected_hash != None:
            ps = '%s/%s.ps' % (wd, test.__name__)
            png = '%s/%s.png' % (wd, test.__name__)
            test(ps)
            os.symlink(png, os.path.join(TEMP_DIR, os.path.basename(png)))
            assert os.path.isfile(png)
            png_hash = sha1(imread(png)).hexdigest()
>           assert png_hash == expected_hash
E           AssertionError: assert 'ac79c75af7ab...f853c40502665' == '903830d2ba5d1...62175bd7b7242'
E             - ac79c75af7ab07f0012ed022fa2f853c40502665
E             + 903830d2ba5d19a415a6d4e15fa62175bd7b7242

tests/test_gmt.py:136: AssertionError
________________________________________________________________________ test_all[__test_cpt-960032a10ef5fef8f013aab8892082fc4a606c9c-5.2] _________________________________________________________________________

test = <function __test_cpt at 0x7fca94beb268>, expected_hash = '960032a10ef5fef8f013aab8892082fc4a606c9c', min_version = 5.2

    @pytest.mark.parametrize("test, expected_hash, min_version", [
        (__test_coastlines, '61efdfbe4cd9bfd5a90ad86c67013c8d9494abc6', 5.0),
        (__test_land, '8c31ef6345aaad068cc4eb3e1d4b3f78cd6fc9a3', 5.0),
        (__test_ticks, 'd625bfb10464664d38f60537686999945eafafe7', 5.0),
        (__test_ticks2, '903830d2ba5d19a415a6d4e15fa62175bd7b7242', 5.0),
        (__test_cpt, '960032a10ef5fef8f013aab8892082fc4a606c9c', 5.2),
        (__test_cpt2, 'd83a3f811e8e514a6e88dacf38fe24bc4dea3280', 5.2),
        (__test_fill, 'eaa069d015eefb20f9826061a6de09d17a420a91', 5.0),
        (__test_autotick, None, 5.0),
        (__test_autotick2, None, 5.0),
        (__test_autotick3, None, 5.0),
    ])
    def test_all(test, expected_hash, min_version):
        wd = os.path.join(TEMP_DIR, test.__name__)
        if not os.path.exists(wd):
            os.makedirs(wd)

        if float('.'.join(gmt.GMT_VERSION.split('.')[:2])) < min_version:
            return

        if expected_hash != None:
            ps = '%s/%s.ps' % (wd, test.__name__)
            png = '%s/%s.png' % (wd, test.__name__)
            test(ps)
            os.symlink(png, os.path.join(TEMP_DIR, os.path.basename(png)))
            assert os.path.isfile(png)
            png_hash = sha1(imread(png)).hexdigest()
>           assert png_hash == expected_hash
E           AssertionError: assert '0c9ce546aa58...bb4f271a6693f' == '960032a10ef5f...082fc4a606c9c'
E             - 0c9ce546aa582abb21d6c3f6837bb4f271a6693f
E             + 960032a10ef5fef8f013aab8892082fc4a606c9c

tests/test_gmt.py:136: AssertionError
________________________________________________________________________ test_all[__test_cpt2-d83a3f811e8e514a6e88dacf38fe24bc4dea3280-5.2] ________________________________________________________________________

test = <function __test_cpt2 at 0x7fca94beb2f0>, expected_hash = 'd83a3f811e8e514a6e88dacf38fe24bc4dea3280', min_version = 5.2

    @pytest.mark.parametrize("test, expected_hash, min_version", [
        (__test_coastlines, '61efdfbe4cd9bfd5a90ad86c67013c8d9494abc6', 5.0),
        (__test_land, '8c31ef6345aaad068cc4eb3e1d4b3f78cd6fc9a3', 5.0),
        (__test_ticks, 'd625bfb10464664d38f60537686999945eafafe7', 5.0),
        (__test_ticks2, '903830d2ba5d19a415a6d4e15fa62175bd7b7242', 5.0),
        (__test_cpt, '960032a10ef5fef8f013aab8892082fc4a606c9c', 5.2),
        (__test_cpt2, 'd83a3f811e8e514a6e88dacf38fe24bc4dea3280', 5.2),
        (__test_fill, 'eaa069d015eefb20f9826061a6de09d17a420a91', 5.0),
        (__test_autotick, None, 5.0),
        (__test_autotick2, None, 5.0),
        (__test_autotick3, None, 5.0),
    ])
    def test_all(test, expected_hash, min_version):
        wd = os.path.join(TEMP_DIR, test.__name__)
        if not os.path.exists(wd):
            os.makedirs(wd)

        if float('.'.join(gmt.GMT_VERSION.split('.')[:2])) < min_version:
            return

        if expected_hash != None:
            ps = '%s/%s.ps' % (wd, test.__name__)
            png = '%s/%s.png' % (wd, test.__name__)
            test(ps)
            os.symlink(png, os.path.join(TEMP_DIR, os.path.basename(png)))
            assert os.path.isfile(png)
            png_hash = sha1(imread(png)).hexdigest()
>           assert png_hash == expected_hash
E           AssertionError: assert '65ab9f5523b4...32d0410350f2e' == 'd83a3f811e8e5...e24bc4dea3280'
E             - 65ab9f5523b4aefbdb15d272f3632d0410350f2e
E             + d83a3f811e8e514a6e88dacf38fe24bc4dea3280

tests/test_gmt.py:136: AssertionError
________________________________________________________________________ test_all[__test_fill-eaa069d015eefb20f9826061a6de09d17a420a91-5.0] ________________________________________________________________________

test = <function __test_fill at 0x7fca94beb378>, expected_hash = 'eaa069d015eefb20f9826061a6de09d17a420a91', min_version = 5.0

    @pytest.mark.parametrize("test, expected_hash, min_version", [
        (__test_coastlines, '61efdfbe4cd9bfd5a90ad86c67013c8d9494abc6', 5.0),
        (__test_land, '8c31ef6345aaad068cc4eb3e1d4b3f78cd6fc9a3', 5.0),
        (__test_ticks, 'd625bfb10464664d38f60537686999945eafafe7', 5.0),
        (__test_ticks2, '903830d2ba5d19a415a6d4e15fa62175bd7b7242', 5.0),
        (__test_cpt, '960032a10ef5fef8f013aab8892082fc4a606c9c', 5.2),
        (__test_cpt2, 'd83a3f811e8e514a6e88dacf38fe24bc4dea3280', 5.2),
        (__test_fill, 'eaa069d015eefb20f9826061a6de09d17a420a91', 5.0),
        (__test_autotick, None, 5.0),
        (__test_autotick2, None, 5.0),
        (__test_autotick3, None, 5.0),
    ])
    def test_all(test, expected_hash, min_version):
        wd = os.path.join(TEMP_DIR, test.__name__)
        if not os.path.exists(wd):
            os.makedirs(wd)

        if float('.'.join(gmt.GMT_VERSION.split('.')[:2])) < min_version:
            return

        if expected_hash != None:
            ps = '%s/%s.ps' % (wd, test.__name__)
            png = '%s/%s.png' % (wd, test.__name__)
            test(ps)
            os.symlink(png, os.path.join(TEMP_DIR, os.path.basename(png)))
            assert os.path.isfile(png)
            png_hash = sha1(imread(png)).hexdigest()
>           assert png_hash == expected_hash
E           AssertionError: assert '54e3ce868fe1...3ce7046de710e' == 'eaa069d015eef...e09d17a420a91'
E             - 54e3ce868fe1a62228f12d91c8a3ce7046de710e
E             + eaa069d015eefb20f9826061a6de09d17a420a91
</pre>
