### 2019-03-05 ucgmsim/GMSimViz#22 @leouieda
Failing XYTS test
The `test_corners` test is failing for me with the following error:

```
_____________________________________ test_corners[True-170.914382935 -43.291595459\n172.262466431 -44.2177429199\n171.012954712 -45.1529998779\n169.66343689 -44.2121429443] ______________________________________

xyts_instance = <gmsimviz.xyts.XYTSFile object at 0x7fcaa7e127f0>, gmt_format = True
expected = '170.914382935 -43.291595459\n172.262466431 -44.2177429199\n171.012954712 -45.1529998779\n169.66343689 -44.2121429443'

    @pytest.mark.parametrize(
        "gmt_format, expected",
        [
            (
                False,
                [
                    [170.914383, -43.291595],
                    [172.262466, -44.217743],
                    [171.012955, -45.153000],
                    [169.663437, -44.212143],
                ],
            ),
            (
                True,
                "170.914382935 -43.291595459\n172.262466431 -44.2177429199\n171.012954712 -45.1529998779\n169.66343689 -44.2121429443",
            ),
        ],
    )
    def test_corners(xyts_instance, gmt_format, expected):
        if gmt_format:
>           assert xyts_instance.corners(gmt_format)[1] == expected
E           AssertionError: assert '170.91438293...1214294433594' == '170.914382935...44.2121429443'
E             - 170.9143829345703 -43.291595458984375
E             ?             - ---             - -----
E             + 170.914382935 -43.291595459
E             - 172.26246643066406 -44.217742919921875
E             ?             ^^^^^^               -----
E             + 172.262466431 -44.2177429199
E             ?             ^...
E
E             ...Full output truncated (8 lines hidden), use '-vv' to show
```

As you can see, the values aren't significantly different and mostly a matter of floating point precision and a different platform. For this reason, it's best to compare floating point numbers within a tolerance. Numpy has convenience functions for that in `numpy.testing` like [numpy.testing.assert_allclose](https://docs.scipy.org/doc/numpy/reference/generated/numpy.testing.assert_allclose.html#numpy.testing.assert_allclose), which I highly recommend.

openjournals/joss-reviews#808
