### 2019-03-05 ucgmsim/GMSimViz#23 @leouieda
Instructions for running the tests
It would be helpful to have some instructions for running the tests (including dependencies). It doesn't have to be in the README for users but maybe in CONTRIBUTING? I managed to run them but only because I read the comment in the JOSS paper review.

openjournals/joss-reviews#808
