### 2019-03-05 #808 @leouieda
@lheagy @viktor76525 sorry for the really long delay. I'm getting caught up now. I went through the changes and am trying to run the tests and examples. I ran into a few issues (ucgmsim/GMSimViz#20 ucgmsim/GMSimViz#21 ucgmsim/GMSimViz#22 ucgmsim/GMSimViz#23) with the tests and missing dependencies. I'm still trying to run the examples and the image comparison but it's taking quite a while. I'll report back when/if the programs finish running.

Overall, I like the changes made to the code itself. It's much easier to find where things are and understand the flow. Nicely done!
