### 2019-03-05 ucgmsim/GMSimViz#20 @viktor76525
fixed in commit: https://github.com/ucgmsim/GMSimViz/commit/72086807e77d5726954e1f85d2084695ae65f77c
