### 2019-03-06 ucgmsim/GMSimViz#22 @viktor76525
Seems even Python 3 was giving different results to Python 2.
Converts string to floats first to compare with tolerance (https://github.com/ucgmsim/GMSimViz/commit/98b4f28842d1f0709d74b911afac6de28023e094)
