### 2019-03-06 c73eb5d
fix issue 23, testing instructions
