+++ CONTRIBUTING.md
@@ -12,5 +12,5 @@ Please use the *Issues* tab on the top of the Github page.
 
 ## Contributing Code
 
-* We do not commit irrelevant files such as `.gitignore`.
+* Tests should be run before contributing code. From the repository root, run: `python -m pytest`.
 * We use [Black](https://github.com/ambv/black) to format the code.
