### 2019-03-06 ucgmsim/GMSimViz#23 @viktor76525
Instructions added (https://github.com/ucgmsim/GMSimViz/commit/c73eb5d976970f50d2b4e2595b6ce55c209a1a77)
