### 2019-03-06 71de7f6
issue 21 improve gmt tests
