+++ tests/test_gmt.py
@@ -1,14 +1,10 @@
 #!/usr/bin/env python2
-"""
-MPI pattern based on
-github.com/jbornschein/mpi4py-examples/blob/master/09-task-pull.py
-"""
 
-from hashlib import sha1
 import os
 from shutil import rmtree
 
 import pytest
+
 try:
     from imageio import imread
 except ImportError:
@@ -16,126 +12,151 @@ except ImportError:
 
 from gmsimviz import gmt
 
-TEMP_DIR = os.path.abspath('GMT_TESTING')
+TEMP_DIR = os.path.abspath("gmt_output")
 if os.path.exists(TEMP_DIR):
     rmtree(TEMP_DIR)
 os.makedirs(TEMP_DIR)
 
-###
-### INTERNAL TESTING FUNCTIONS
-###
 
-def __test_coastlines(pf):
-    p = gmt.GMTPlot(pf)
-    p.spacial('M', (170, 175, -44, -38), sizing = 3)
-    p.coastlines(width = '0.4p')
-    p.finalise()
-    p.png(dpi = 200, clip = True)
+class Figure:
+    """
+    Creates an object with a savefig function for pytest-mpl.
+    """
 
-def __test_land(pf):
-    p = gmt.GMTPlot(pf)
-    p.spacial('M', (170.1, 179.91, -37, -34), sizing = 7)
-    p.land(fill = 'darkred')
-    p.finalise()
-    p.png(dpi = 100, clip = True)
+    # to make matplotlib.pyplot.close(Figure) work
+    int = None
 
-def __test_ticks(pf):
-    p = gmt.GMTPlot(pf)
-    p.spacial('M', (160.992, 174.9122, -44, -34.01), sizing = 2)
-    p.ticks(major = '1d', minor = '20m', sides = 'ew')
-    p.finalise()
-    p.png(dpi = 100, clip = True)
+    def __init__(self, gmt_plot, dpi=100, clip=True):
+        self.p = gmt_plot
+        self.dpi = dpi
+        self.clip = clip
 
-def __test_ticks2(pf):
-    p = gmt.GMTPlot(pf)
-    p.spacial('T', (160.992, 174.9122, -44, -34.01), sizing = 5, \
-            x_shift = 0.5, y_shift = 0.5, lon0 = 174.9122)
-    p.ticks(major = '1d', minor = '20m', sides = 'ew')
-    p.finalise()
-    p.png(dpi = 100, clip = True)
+    def savefig(self, filename):
+        self.p.finalise()
+        self.p.png(out_name=os.path.splitext(filename)[0], dpi=self.dpi, clip=self.clip)
+        rmtree(self.p.wd)
 
-def __test_cpt(pf):
-    cptf = '%s.cpt' % (os.path.splitext(pf)[0])
-    gmt.makecpt('hot', cptf, 0, 120, inc = 0.1, invert = True, \
-            wd = os.path.dirname(pf))
-    p = gmt.GMTPlot(pf)
-    p.spacial('X', (0, 15, 0, 4), sizing = '15/4')
-    p.cpt_scale(6.1, 2.05, cptf, 20, 5, label = 'test_scale', \
-            length = 3.05, thickness = '0.3i')
-    p.finalise()
-    p.png(dpi = 320, clip = True)
 
-def __test_cpt2(pf):
-    cptf = '%s.cpt' % (os.path.splitext(pf)[0])
-    gmt.makecpt('polar', cptf, -1.5, 1.5, inc = 0.25, invert = False, \
-            wd = os.path.dirname(pf), bg = '0/0/80', fg = '80/0/0')
-    p = gmt.GMTPlot(pf)
-    p.spacial('X', (0, 4, 0, 2), sizing = '4/2', x_shift = 1, y_shift = 1)
[…]
