### 2019-03-06 ucgmsim/GMSimViz#21 @viktor76525
fixed everything up here (https://github.com/ucgmsim/GMSimViz/commit/71de7f639a8228f208c107f08815a36e5bb73745)
tests are now run with `py.test --mpl` from the tests directory (documentation updated)
