### 2019-03-06 #808 @viktor76525
@leouieda Thanks for the pytest-mpl suggestion. I have closed all issues and ready for a (hopefully) quick look.
