### 2019-03-07 ucgmsim/GMSimViz#22 @leouieda
:+1: I can confirm that these tests now pass for me :tada:

@viktor76525 not necessary right now but a tip when using `allclose`: if you do `assert np.allclose(...)` you won't get any information when it fails. But `np.assert_allclose(...)` tells you how different the two values are and what the actual values are.
