### 2019-03-07 ucgmsim/GMSimViz#21 @leouieda
:+1: these tests pass for me with GMT 5.4.4 and ghostscript 9.20. With newer GMT/ghostscript there are some failures because of slight changes in the images (not significant). This is a bit of a problem with testing images (a lot of false negatives) but I  don't know of a good way around it.
