### 2019-03-07 ucgmsim/GMSimViz#23 @leouieda
@viktor76525 what about the `tests/gmsimviz_compare.py` script? Is this part of the test suite?
