### 2019-03-07 ucgmsim/GMSimViz#24 @leouieda
Errors and "squished" image when running example
I tried running the first example in the README (`gmsimviz sample_data/fault.srf` from the root of the repository) and got the expected error messages but also some other ones complaining about files not found:

```
psxy: Error for input file: No such file (resources/Paths/lds-nz-coastlines-and-islands/150k.gmt)
grdimage: Error: (x_max-x_min) must equal (NX + eps) * x_inc), where NX is an integer and |eps| <= 0.0001.
grdimage (gmtlib_read_grd_info): Use grdedit -A on your grid file to make region and increments compatible [/home/leo/src/GMSimViz/resources/Topo/srtm_all_filt_nz_i5.grd]
grdimage: Error: (x_max-x_min) must equal (NX + eps) * x_inc), where NX is an integer and |eps| <= 0.0001.
grdimage (gmtlib_read_grd_info): Use grdedit -A on your grid file to make region and increments compatible [/home/leo/src/GMSimViz/resources/Topo/srtm_all_filt_nz.grd]
grdimage: File /home/leo/miniconda3/envs/GMSymVis/share/gmt/cpt/resources/cpt/palm_springs_nz_topo.cpt not found
[Session gmt (0)]: Error returned from GMT API: GMT_FILE_NOT_FOUND (16)
psclip: Error for input file: No such file (resources/Paths/lds-nz-coastlines-and-islands/150k.gmt)
psxy: Error for input file: No such file (resources/Paths/lds-nz-lake-polygons/150k.gmt)
psxy: Error for input file: No such file (resources/Paths/lds-nz-river-polygons/150k.gmt)
psxy: Error for input file: No such file (resources/Paths/lds-nz-coastlines-and-islands/150k.gmt)
```

The files do exist in `resources`. But the I don't know why it's looking for the `palm_springs_nz_topo.cpt` file in my GMT install directory instead of the `resources` folder.

The resulting image is also distorted with a lot of empty transparent space (`fault_perspective.png`):

![fault_perspective](https://user-images.githubusercontent.com/290082/53979160-1c517400-40b1-11e9-942c-514d7de7b5c3.png)

openjournals/joss-reviews#808
