### 2019-03-07 #808 @leouieda
@viktor76525 I went through most of the changes and I can get the tests to run now with pytest-mpl. What about the `gmsimviz_compare.py` file? Is that not part of the test suite?

I'm trying to run the examples but I'm getting some errors (see ucgmsim/GMSimViz#24).
