### 2019-03-07 ucgmsim/GMSimViz#25 @leouieda
In the examples, explain which files should be generated and where
The examples in the README contain the commands that should be run and the expected image/video. But there is nothing said about the files that are generated in the process or their names. One could guess from the files that suddenly pop up in the directory but it would be nice to have some mention of that in the text as well so we can be certain that we got the right output.

openjournals/joss-reviews#808
