### 2019-03-07 ucgmsim/GMSimViz#24 @viktor76525
Did you get any error when running setup.py `python setup.py install --user` or similar?
The reason is the missing paths are relative when they have to be absolute.
"resources" gets changed to the absolute path (in gmsimviz/config.json) at the start of setup.py. the original is just to make tests work before the gmsimviz package is installed

`/home/vap30/.local/lib64/python3.6/site-packages/GMSimViz-1.2.1-py3.6.egg/gmsimviz/config.json`
where your path will be slightly different should contain an absolute path:
```
{
"GMT_DATA" : "/home/vap30/ucgmsim/GMSimViz/resources"
}
```

I am not sure about the squashed image but it could be a side effect of other errors.
