### 2019-03-07 2d3c638
py.test will work only if gmsimviz is already installed without extra parameters anyway
