### 2019-03-07 41db4fb
added note to run non-unit testing code
