### 2019-03-08 ucgmsim/GMSimViz#23 @viktor76525
gmsimviz_compare.py is a separate non unit-test test that runs an animation, takes a few images out and compares the amount of pixels changed
