### 2019-03-08 ucgmsim/GMSimViz#25 @viktor76525
fixed (https://github.com/ucgmsim/GMSimViz/commit/2266d00a5d375a0c9d2bb011bc54de34524ec79c)
