### 2019-03-08 ucgmsim/GMSimViz#24 @leouieda
OK, I managed to track it down. I had installed the project with `pip install .` to test it and forgot to revert back to `setup.py install`. Doing the `python setup.py install` got the example to run correctly. So something in your `setup.py` makes it incompatible with pip, which might be worth fixing since pip is the defacto way of installing Python software.
