### 2019-03-08 ucgmsim/GMSimViz#24 @leouieda
To be clear, I don't think making it compatible with pip is a requirement for the JOSS publication. It's just a recommendation that will make GMSimViz easier to use. The instructions on the README work, they're just not following the Python community standards.
