### 2019-03-08 ucgmsim/GMSimViz#24 @viktor76525
I wasn't aware of using pip to install packages in that way.
This is getting tricky. I can override the install function to add the absolute path first but pip will first copy setup.py to a temporary location so the path is wrong (can't use __file__ to find repo root). Ideally resources shouldn't be part of the package data because of their size and being less-related to code, updates in one vs the other.
The other option is make the resource folder location a non-optional parameter to the install but I was trying to streamline the install process.
