### 2019-03-08 ucgmsim/GMSimViz#24 @leouieda
The advantage of having it pip-compatible is that you could then publish to PyPI and allow people to `pip install gmsimviz`. But hosting a lot of data is a difficult thing. I hadn't considered that. How about having the resources be defined at run-time?

If they're "not part of the package" then the user should be able to control where they are stored. That's how I handle sample data for my packages. I built a library called [pooch](https://github.com/fatiando/pooch) that takes care of downloading the data and saving it somewhere. It has built-in support for checking user defined environment variables.
