### 2019-03-08 #808 @leouieda
@lheagy @viktor76525 I've gone through the submission one more time and managed to install, get all tests passing, and run the examples. The code has improved a lot since submission and is following a much more standard Python project layout (besides some minor issues with making it pip-installable which we're discussing in ucgmsim/GMSimViz#24).

I don't have any more comments to add and to me the submission ticks all the required boxes.
