### 2019-03-08 ucgmsim/GMSimViz#24 @viktor76525
I think this covers it (https://github.com/ucgmsim/GMSimViz/commit/6e831ce638dedb93fbeaf893d4f8d07a853e07fd)
