### 2019-03-08 ucgmsim/GMSimViz#24 @leouieda
In this case, the `setup.py` is still trying to set the resources right?
