### 2019-03-08 ucgmsim/GMSimViz#24 @viktor76525
only if run directly, not with pip. do you think it should be optional either way?
