### 2019-03-08 ucgmsim/GMSimViz#24 @leouieda
No, that's fine since the user can control it at runtime with the environment variable. Nice!
