### 2019-03-17 #808 @lheagy
:wave: Hi @openjournals/joss-eics, this submission is ready to be accepted 🎉
