+++ paper/paper.bib
@@ -1,6 +1,7 @@
 @article{gmt,
 author = {Wessel, Paul and Smith, Walter H. F. and Scharroo, Remko and Luis, Joaquim and Wobbe, Florian},
 title = {Generic Mapping Tools: Improved Version Released},
+year = {2013},
 journal = {Eos, Transactions American Geophysical Union},
 volume = {94},
 number = {45},
