### 2019-03-18 #808 @labarba
Congratulations, @viktor76525, your JOSS paper is published!

Sincere thanks to the editor: @lheagy, and the reviewers: @leouieda, @hugoledoux 🙏  We couldn't do this without you!
